Moving structs where they belong
