+++ FFXIVClientStructs/FFXIV/Client/System/Input/TextServiceInterface.cs
@@ -1,0 +1,12 @@
+namespace FFXIVClientStructs.FFXIV.Client.System.Input;
+
+// Client::System::Input::TextServiceInterface
+[GenerateInterop(isInherited: true)]
+[StructLayout(LayoutKind.Explicit, Size = 0x8)]
+public partial struct TextServiceInterface {
+
+    // Client::System::Input::TextServiceInterface::TextServiceEvent
+    [GenerateInterop(isInherited: true)]
+    [StructLayout(LayoutKind.Explicit, Size = 0x8)]
+    public partial struct TextServiceEvent;
+}
