+++ FFXIVClientStructs/FFXIV/Client/UI/Misc/RetainerCommentModule.cs
@@ -1,4 +1,4 @@
-using FFXIVClientStructs.FFXIV.Client.UI.Misc.UserFileManager;
+using UserFileEvent = FFXIVClientStructs.FFXIV.Client.UI.Misc.UserFileManager.UserFileEvent;
 
 namespace FFXIVClientStructs.FFXIV.Client.UI.Misc;
